Lucylovers/requirements.txt
torch
torchvision
torch.nn
torch.optim
matplotlib.pyplot
torch.autograd
torchvision.transforms
sklearn.metrics
os
torch.nn.functional